+++ src/CommandingOfficers/CommanderLibrary.java
@@ -5,6 +5,8 @@ import java.util.ArrayList;
 import CommandingOfficers.AW4.BrennerWolves.*;
 import CommandingOfficers.DefendPeace.CyanOcean.Ave;
 import CommandingOfficers.DefendPeace.CyanOcean.Patch;
+import CommandingOfficers.DefendPeace.RoseThorn.Strong;
+import CommandingOfficers.DefendPeace.RoseThorn.Tech;
 import CommandingOfficers.DefendPeace.misc.Bear_Bull;
 import CommandingOfficers.DefendPeace.misc.DocLight;
 import CommandingOfficers.DefendPeace.misc.Meridian;
