+++ src/CommandingOfficers/DefendPeace/RoseThorn/Strong.java
@@ -1,7 +1,12 @@
-package CommandingOfficers;
+package CommandingOfficers.DefendPeace.RoseThorn;
 
 import java.util.ArrayList;
 
+import CommandingOfficers.Commander;
+import CommandingOfficers.CommanderAbility;
+import CommandingOfficers.CommanderInfo;
+import CommandingOfficers.CommanderAbility.CostBasis;
+import CommandingOfficers.CommanderInfo.InfoPage;
 import CommandingOfficers.Modifiers.UnitProductionModifier;
 import Engine.GameScenario;
 import Engine.Combat.StrikeParams.BattleParams;
