+++ src/Test/TestAIBehavior.java
@@ -5,8 +5,8 @@ import AI.AIMaker;
 import AI.Muriel;
 import AI.WallyAI;
 import CommandingOfficers.Commander;
-import CommandingOfficers.Strong;
 import CommandingOfficers.DefendPeace.CyanOcean.Patch;
+import CommandingOfficers.DefendPeace.RoseThorn.Strong;
 import Engine.Army;
 import Engine.GameAction;
 import Engine.GameInstance;
