+++ src/Test/TestCombat.java
@@ -1,8 +1,8 @@
 package Test;
 
 import CommandingOfficers.Commander;
-import CommandingOfficers.Strong;
 import CommandingOfficers.DefendPeace.CyanOcean.Patch;
+import CommandingOfficers.DefendPeace.RoseThorn.Strong;
 import Engine.Army;
 import Engine.GameAction;
 import Engine.GameInstance;
